remove ButtonWithIcon
