+++ src/componentAtoms/index.js
@@ -48,7 +48,7 @@ import { HeaderCollapsedPropedNavbarIndex } from './HeaderCollapsed'
 import { DropdownMenuNavbarStoreUserPropedNavbar } from './DropdownMenuNavbar'
 import { AvatarUserStoreUser } from './AvatarUser'
 import { DropdownNavbarStoreUser } from './DropdownNavbar'
-import { ButtonGroupProfile } from './ButtonWithIcon'
+import { ButtonGroupProfile } from './ButtonsIcon'
 import { QuantityProfile } from './Quantity'
 import {
 	CarouselCommonPropedProfile,
